==> 1_get_covers.ipynb <==
import numpy as np
import requests
from bs4 import BeautifulSoup
import pandas as pd
from tqdm import tqdm
import glob
import os
from PIL import Image
import re

BASE_URL = "https://naruto.fandom.com"
volume_list_url = "/wiki/List_of_Volumes"
res = requests.get(BASE_URL + volume_list_url)
soup = BeautifulSoup(res.text, "html.parser")
tables = soup.find_all("table", {"class": "wikitable"})
# The main Naruto series has 72 volumes, published in two parts of 27 and 45 volumes, respectively. These volumes are in the first two tables on the website.
volumes = tables[0].find_all("div", class_="floatright") + tables[1].find_all(
    "div", class_="floatright"
)
volume_urls = [volume.find("a").get("href") for volume in volumes]

volumes = {}

for volume, volume_url in tqdm(enumerate(volume_urls)):
    res = requests.get(BASE_URL + volume_url)
    soup = BeautifulSoup(res.text, "html.parser")
    infobox = soup.find("table", class_="infobox")
    volumes[volume] = {
        "name": infobox.find("th", class_="mainheader").find("i").text,
        "number": int(
            infobox.find("th", class_="", string=re.compile(r"Volume"))
            .find_next_sibling("td")
            .find(string=True)
            .strip()
        ),
        "release_date": (
            infobox.find("th", class_="", string=re.compile(r"English"))
            .find_next_sibling("td")
            .find(string=True)
            .strip()
        ),
        "cover_url": infobox.find("a", class_="image").get("href"),
    }

volumes_df = pd.DataFrame.from_dict(volumes, orient="index").set_index("number")
volumes_df["release_date"] = pd.to_datetime(volumes_df["release_date"])
volumes_df.to_csv("data/volumes.csv")

Download all the covers

for i, cover_url in tqdm(enumerate(volumes_df.cover_url.values)):
    res = requests.get(cover_url)
    with open("./data/covers/{}.webp".format(i + 1), "wb") as f:
        f.write(res.content)

sizes = {}
for file in tqdm(glob.glob("./data/covers/*.jpg")):
    img = Image.open(file)
    if img.size not in sizes:
        sizes[img.size] = 1
    else:
        sizes[img.size] += 1
# Sort by most common size
sizes = sorted(sizes.items(), key=lambda x: x[1], reverse=True)
sizes

Most of the images seem to be around (760, 1200) in size. So lets resize all the images to be that size.

images = glob.glob("data/covers/*")
for image in tqdm(images):
    img = Image.open(image)
    img = img.resize((760, 1200))
    img.save("data/small_covers/" + os.path.basename(image))
