==> 2_large_list_of_names.ipynb <==
import io
import requests

import pandas as pd

Get the Large List of Named numbers from http://mkweb.bcgsc.ca/colornames/color.names.txt

req = requests.get("http://mkweb.bcgsc.ca/colornames/color.names.txt")
# Remove all lines that start with a '#'. These are comments in the file
comments = [line for line in req.text.split("\n") if line.startswith("#")]
color_lines = [line for line in req.text.split("\n") if not line.startswith("#")]
# Create a dataframe with the color names
colors = pd.read_csv(
    io.StringIO("\n".join(color_lines)), sep=" ", header=None, index_col=0
)

print("\n".join(comments))

# Using the comments above, we can create a header row for the dataframe
header = [
    "name",
    "rgb_head",
    "rgb_R",
    "rgb_G",
    "rgb_B",
    "hex_head",
    "hex",
    "hsv_head",
    "hsv_H",
    "hsv_S",
    "hsv_V",
    "xyz_head",
    "xyz_X",
    "xyz_Y",
    "xyz_Z",
    "lab_head",
    "lab_L",
    "lab_a",
    "lab_b",
    "lch_head",
    "lch_L",
    "lch_C",
    "lch_h",
    "cmyk_head",
    "cmyk_C",
    "cmyk_M",
    "cmyk_Y",
    "cmyk_K",
    "nearest_neighbours",
    "num_nearest_neighbours_within_5",
    "num_nearest_neighbours_within_5_names",
]
colors.columns = header

# All columns which have the same value for all rows
columns_to_drop = colors.columns[colors.nunique() == 1]
colors = colors.drop(columns=columns_to_drop)

colors.to_csv("./data/colors.names.csv", index=False)
